Add tests for withIon
