+++ tests/utils/waitForPromisesToResolve.js
@@ -1,0 +1,1 @@
+export default () => new Promise(setImmediate);
